Implemented the context driven input system.
The input system monitors for user input and triggers the corresponding events. The action taken based on the
event depends on the current context being used. The contexts are dynamically swappable. This allows for
different uses for the same user input depending on the context. i.e. gameplay vs menu naviation.

Demo Scene is included in Assets/Scenes/InputSystem/InputSystemDemo.unity